Add support for one-time URL flags
A one-time flag (OTF) is a URL query parameter that is added to only one
HTTP request made with a LrHandle instance.  Any subsequent requests
with the same handle, including any retry attempts made by librepo
internally, do not carry the OTF.

This property is useful for implementing a mechanism for counting the
users of a repository (and, by extension, of the OS as a whole) in which
all OTF instances received by the server in a specific time window are
summed up to produce a good estimate, with the advantage of bundling the
flag into regular metadata updates and not sending separate requests
just for that purpose alone.

The option is called LRO_ONETIMEFLAG and is a key=value string.

The guarantee is that, once set on a LrHandle, the OTF will only be used
once and then thrown away (set to NULL).  It is up to the program to
implement a time window logic that sets the option repeatedly (such as
once per week) to achieve the desired effect.

Note that the OTF is only applied to the metalink and mirrorlist URL
(whichever is fetched first) as those are the common entry points for
most OS-level repositories.

For a reference implementation, see the "countme" option in dnf.conf(5)
or the associated Fedora change:

https://fedoraproject.org/wiki/Changes/DNF_Better_Counting

Technical notes:

- OTF is implemented as a per-target property since we only want to
  apply it to specific metadata downloads (metalink and mirrorlist) as
  opposed to all downloads, which is why we define
  LrDownloadTarget.onetimeflag and make it inherit the value from its
  parent handle.

- The signature of the internal function lr_yum_download_url() is being
  changed.